move of letting view and url into a dedicated app
